chore: move calculation by consumption to its own page
