+++ src/main.tsx
@@ -6,8 +6,9 @@ import './global.css'
 import { DashboardLayout } from './app/index.tsx';
 import { DailyChallengePage } from './app/dailyChallenge/index.tsx';
 import { ReminderPage } from './app/reminder/index.tsx';
-import { CalculatorByDevice, CalculatorByPrice } from './app/energyCalculator/index.tsx';
+import { CalculatorByDevice} from './app/energyCalculator/index.tsx';
 import { CalculatorLayout } from './app/energyCalculator/layout.tsx';
+import { CalculatorByPrice } from './app/energyCalculator/byPrice.tsx';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
